+++ xbmc/cores/VideoPlayer/DVDDemuxers/DVDDemuxBXA.h
@@ -61,7 +61,7 @@ public:
   void Flush() override;
   DemuxPacket* Read() override;
   bool SeekTime(double time, bool backwards = false, double* startpts = NULL) override { return false; }
-  int GetStreamLength() override { return (int)m_header.durationMs; }
+  int64_t GetStreamLength() override { return (int)m_header.durationMs; }
   CDemuxStream* GetStream(int iStreamId) const override;
   std::vector<CDemuxStream*> GetStreams() const override;
   int GetNrOfStreams() const override;
