+++ xbmc/cores/VideoPlayer/DVDDemuxers/DVDDemuxCDDA.h
@@ -42,7 +42,7 @@ public:
   void Flush() override;
   DemuxPacket* Read() override;
   bool SeekTime(double time, bool backwards = false, double* startpts = NULL) override;
-  int GetStreamLength() override;
+  int64_t GetStreamLength() override;
   CDemuxStream* GetStream(int iStreamId) const override;
   std::vector<CDemuxStream*> GetStreams() const override;
   int GetNrOfStreams() const override;
